+++ bocadillo/__init__.py
@@ -9,4 +9,4 @@ from .websockets import WebSocket, WebSocketDisconnect
 from .response import Response
 from .request import Request
 
-__version__ = "0.11.0"
+__version__ = "0.11.1"